+++ website/src/layout.rs
@@ -34,6 +34,7 @@ pub fn Layout(children: Children) -> impl IntoView {
                 <BNavbarStart>
                     <BNavbarItem href="/">{t!(i18n, home)}</BNavbarItem>
                     <BNavbarItem href="/guides">{t!(i18n, guides)}</BNavbarItem>
+                    <BNavbarItem href="/components">{t!(i18n, components)}</BNavbarItem>
                 </BNavbarStart>
 
                 <BNavbarEnd>