+++ website/src/pages/mod.rs
@@ -1,7 +1,9 @@
+mod components_page;
 mod guides_page;
 mod home_page;
 mod not_found_page;
 
+pub use components_page::ComponentsPage;
 pub use guides_page::GuidesPage;
 pub use home_page::HomePage;
 pub use not_found_page::NotFoundPage;